[jvm] Introduce the Address REPR
Socket addresses are currently represented with host/port pairs. This is
only adequate for IPv4 addresses: IPv6 addresses may have a network
interface associated with them, and whenever the time comes for this to
apply, UNIX socket addresses do not carry ports.

Socket addresses are to be represented with the Address REPR, which
stores a socket address wrapped by one of the new NetworkAddress
subclasses.