[jvm] Add IPAddress v5.3.3 as a dependency
For some godforsaken reason, there is no way to construct a
java.net.Inet4Address or java.net.Inet6Address with IP literals without
the risk of DNS getting involved without relying on JDK internals (which
doesn't play nice with javac --release). The upcoming addrfromstr_ip4
and addrfromstr_ip6 ops depend on this being possible. IPAddress is a
fairly minimal library that can take care of this for us, as parsing IP
addresses isn't as trivial a task as it may seem.