[jvm] Stub the addrfromstr_un op
The JVM backend has no reasonable means of implementing UNIX socket
support at this point in time (JDK 16 will change things in that
regard). We'd need this to implement the addrfromstr_un op, which
creates a UNIX socket address from a path, so just throw a NYI exception
for now.