[jvm] Implement the dnslookup op
This performs a getaddrinfo(3)-like DNS resolution given a hostname,
port, protocol family, socket type, protocol type, and flags. Currently,
the AI_ADDRCONFIG and AI_PASSIVE flags are supported.

This sets the java.net.preferIPv4Stack property to false. This is the
default, but with OpenBSD's OpenJDK, this is configured to be true
instead, preventing IPv6 addresses from being resolved.